+++ test/unit/specs/DatePicker.spec.js
@@ -5,17 +5,13 @@ describe('DatePicker', () => {
   it('renders props.fromPage', () => {
     const fromPage = { month: 4, year: 2001 };
     const wrapper = mount(DatePicker, {
-      attrs: {
+      propsData: {
         fromPage,
         isInline: true,
       },
     });
     const cal = wrapper.find('div');
     console.log(cal.html());
-    // if (cal) {
-    //   const attrs = cal.attributes();
-    //   console.log(attrs);
-    // }
     expect(2).toBe(2);
   });
 });